backend/pyproject.toml
[project]
name = "oneshot-backend"
version = "0.1.0"
description = "oneshot backend"
authors = [
    {name = "mungnpang",email = "koressma@naver.com"}
]
requires-python = "^3.11"
dependencies = [
    "fastapi (>=0.135.2,<0.136.0)",
    "uvicorn[standard] (>=0.42.0,<0.43.0)",
    "sqlalchemy (>=2.0.48,<3.0.0)",
    "alembic (>=1.18.4,<2.0.0)",
    "pydantic-settings (>=2.13.1,<3.0.0)",
    "python-jose[cryptography] (>=3.5.0,<4.0.0)",
    "python-multipart (>=0.0.22,<0.0.23)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "bcrypt (<5.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "rapidfuzz (>=3.0.0)",
    "google-genai (>=1.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest (>=9.0.2,<10.0.0)",
    "pytest-asyncio (>=1.3.0,<2.0.0)"
]
